:hammer: Refactor: 리팩터링 마무리 및 noImplicitAny 적용
