+++ backend/src/graphql/services/PostService.ts
@@ -5,16 +5,19 @@ import { PostQuestionHasTag } from "../../entities/PostQuestionHasTag";
 import { Tag } from "../../entities/Tag";
 import { User } from "../../entities/User";
 import AddQuestionInput from "../inputTypes/AddQuestionInput";
+import SearchQuestionInput from "../inputTypes/SearchQuestionInput";
 
 export default class PostService {
   private static DEFALUT_TAKE_QUESTIONS_COUNT = 20;
-  public static async findAllQuestionByArgs(args): Promise<PostQuestion[]> {
+  public static async findAllQuestionByArgs(
+    args: SearchQuestionInput
+  ): Promise<PostQuestion[]> {
     const { author, tagIDs, skip, take } = args;
-    const { title, desc, realtime_share } = args;
+    const { title, desc, realtimeShare } = args;
 
     const whereObj: Record<string, unknown> = {};
 
-    if (realtime_share) whereObj.realtimeShare = realtime_share;
+    if (realtimeShare) whereObj.realtimeShare = realtimeShare;
     if (title) whereObj.title = Like(`%${title}%`);
     if (desc) whereObj.desc = Like(`%${desc}%`);
 
@@ -66,8 +69,18 @@ export default class PostService {
     return await builder.getMany();
   }
 
-  public static async findAllAnswerByArgs(args): Promise<PostAnswer[]> {
-    const data = await PostAnswer.find(args);
+  public static async findAllQuestionByUserId(
+    userId: number
+  ): Promise<PostQuestion[]> {
+    const questions = await PostQuestion.find({ userId });
+
+    return questions;
+  }
+
+  public static async findAllAnswerByUserId(
+    userId: number
+  ): Promise<PostAnswer[]> {
+    const data = await PostAnswer.find({ userId });
 
     return data;
   }
@@ -80,7 +93,7 @@ export default class PostService {
     return data;
   }
 
-  public static async findOneQuestionById(id): Promise<PostQuestion> {
+  public static async findOneQuestionById(id: number): Promise<PostQuestion> {
     const question = await PostQuestion.findOne({ id: id });
 
     return question;
@@ -88,7 +101,8 @@ export default class PostService {
 
   public static async addNewQuestion(
     args: AddQuestionInput,
-    user
+    // 이후 ctx.user 로 수정
+    user: { id: number }
   ): Promise<PostQuestion> {
     const newQuestion = new PostQuestion();
     newQuestion.userId = user.id;
@@ -107,17 +121,5 @@ export default class PostService {
     }
 
     return newQuestion;
-  }
-
-  public static async getAllTagIdsByQuestionId(id: number): Promise<number[]> {
-    const question = await PostQuestion.find({ id: id });
-    const tagRelations = await createQueryBuilder()
-      .relation(PostQuestion, "postQuestionHasTags")
-      .of(question)
-      .loadMany();
-
-    const tagIds = tagRelations.map((obj) => obj.tagId);
-
-    return tagIds;
   }
 }
