+++ backend/src/graphql/services/TagService.ts
@@ -1,4 +1,6 @@
 import { Tag } from "../../entities/Tag";
+import { PostQuestion } from "../../entities/PostQuestion";
+import { createQueryBuilder } from "typeorm";
 
 export default class TagService {
   public static async getAllTags(): Promise<Tag[]> {
@@ -17,5 +19,17 @@ export default class TagService {
 
   public static async findTagByName(name: string): Promise<Tag> {
     return Tag.findOne({ name });
+  }
+
+  public static async getAllTagIdsByQuestionId(id: number): Promise<number[]> {
+    const question = await PostQuestion.find({ id: id });
+    const tagRelations = await createQueryBuilder()
+      .relation(PostQuestion, "postQuestionHasTags")
+      .of(question)
+      .loadMany();
+
+    const tagIds = tagRelations.map((obj) => obj.tagId);
+
+    return tagIds;
   }
 }
